Refactored file locations, refactored imports
