+++ app/src/main/java/com/example/starca/models/Listing.kt
@@ -1,7 +1,7 @@
 package com.example.starca.models
 
 import android.util.Log
-import com.example.starca.ListingRequest
+import com.example.starca.models.ListingRequest
 import com.parse.ParseClassName
 import com.parse.ParseFile
 import com.parse.ParseObject
